rename to ostream
