+++ exporters/CMakeLists.txt
@@ -1,4 +1,4 @@
 if(WITH_OTPROTOCOL)
   add_subdirectory(otlp)
 endif()
-add_subdirectory(stream)
+add_subdirectory(ostream)
