+++ exporters/ostream/BUILD
@@ -1,12 +1,12 @@
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "stream_span_exporter",
+    name = "ostream_span_exporter",
     srcs = [
         "src/span_exporter.cc",
     ],
     hdrs = [
-        "include/opentelemetry/exporters/stream/span_exporter.h",
+        "include/opentelemetry/exporters/ostream/span_exporter.h",
     ],
     strip_include_prefix = "include",
     deps = [
@@ -15,10 +15,10 @@ cc_library(
 )
 
 cc_test(
-    name = "stream_span_test",
-    srcs = ["test/stream_span_test.cc"],
+    name = "ostream_span_test",
+    srcs = ["test/ostream_span_test.cc"],
     deps = [
-        ":stream_span_exporter",
+        ":ostream_span_exporter",
         "@com_google_googletest//:gtest_main",
     ],
 )
